+++ src/main/scala/sigmastate/lang/Terms.scala
@@ -8,7 +8,7 @@ import sigmastate.interpreter.Context
 object Terms {
 
   case class Block(bindings: Seq[Let], result: SValue) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = false
     def tpe: SType = result.tpe
   }
@@ -17,7 +17,7 @@ object Terms {
   }
 
   case class Let(name: String, givenType: SType, body: SValue) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = ???
     def tpe: SType = givenType ?: body.tpe
   }
@@ -26,7 +26,7 @@ object Terms {
   }
 
   case class Select(obj: Value[SType], field: String, resType: Option[SType] = None) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = ???
     val tpe: SType = resType.getOrElse(obj.tpe match {
       case p: SProduct =>
@@ -38,7 +38,7 @@ object Terms {
   }
 
   case class Ident(name: String, tpe: SType = NoType) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = ???
   }
   object Ident {
@@ -46,7 +46,7 @@ object Terms {
   }
 
   case class Apply(func: Value[SType], args: IndexedSeq[Value[SType]]) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = false
     lazy val tpe: SType = func.tpe match {
       case SFunc(_, r, _) => r
@@ -57,7 +57,7 @@ object Terms {
 
   /** Apply types for type parameters of input value. */
   case class ApplyTypes(input: Value[SType], tpeArgs: Seq[SType]) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = false
     lazy val tpe: SType = input.tpe match {
       case funcType: SFunc =>
@@ -68,12 +68,12 @@ object Terms {
   }
 
   case class MethodCall(obj: Value[SType], name: String, args: IndexedSeq[Value[SType]], tpe: SType = NoType) extends Value[SType] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = false
   }
 
   case class Lambda(args: IndexedSeq[(String,SType)], givenResType: SType, body: Option[Value[SType]]) extends Value[SFunc] {
-    override def cost[C <: Context[C]](context: C): Int = ???
+    override def cost[C <: Context[C]](context: C): Long = ???
     override def evaluated: Boolean = false
     lazy val tpe: SFunc = SFunc(args.map(_._2), givenResType ?: body.fold(NoType: SType)(_.tpe))
   }
